requirements.txt
aiohttp>=3.8.0
shodan>=1.28.0
requests>=2.28.0
typing-extensions>=4.5.0
python-dateutil>=2.8.2
asyncio>=3.4.3
python-whois>=0.8.0
dnspython>=2.3.0
tld>=0.13
idna>=3.4
chardet>=5.1.0
pyOpenSSL>=23.2.0
cryptography>=41.0.5
OTXv2>=1.5.12
tqdm>=4.65.0
ipaddress>=1.0.23
argparse>=1.4.0
certifi>=2023.7.22
urllib3>=2.0.0
